refactor(noncomp): align sampling and status walk boundaries
