+++ src/clifft/noncomp/exact_driver.cc
@@ -46,10 +46,9 @@
 #include "clifft/frontend/frontend.h"
 #include "clifft/noncomp/annotate.h"
 #include "clifft/noncomp/instrument_options.h"
-#include "clifft/noncomp/op_role.h"
 #include "clifft/noncomp/rewriter.h"
 #include "clifft/noncomp/seed.h"
-#include "clifft/noncomp/status_step.h"
+#include "clifft/noncomp/status_walk.h"
 #include "clifft/noncomp/transition_instrument.h"
 #include "clifft/optimizer/hir_pass_manager.h"
 #include "clifft/optimizer/pass_registry.h"
