+++ src/clifft/noncomp/rewriter.cc
@@ -5,8 +5,7 @@
 #include "clifft/noncomp/classifier.h"
 #include "clifft/noncomp/level.h"
 #include "clifft/noncomp/numeric.h"
-#include "clifft/noncomp/op_role.h"
-#include "clifft/noncomp/status_step.h"
+#include "clifft/noncomp/status_walk.h"
 #include "clifft/noncomp/transition_instrument.h"
 
 #include <algorithm>
