+++ src/python/bindings.cc
@@ -5,8 +5,8 @@
 #include "clifft/frontend/frontend.h"
 #include "clifft/noncomp/level.h"
 #include "clifft/noncomp/model.h"
-#include "clifft/noncomp/orchestrator.h"
 #include "clifft/noncomp/policy.h"
+#include "clifft/noncomp/sample.h"
 #include "clifft/optimizer/bytecode_pass.h"
 #include "clifft/optimizer/drop_non_unitary_pass.h"
 #include "clifft/optimizer/expand_t_pass.h"
