+++ tests/CMakeLists.txt
@@ -41,8 +41,7 @@ add_executable(clifft_tests
     test_noncomp_transition_instrument.cc
     test_noncomp_classifier.cc
     test_noncomp_model.cc
-    test_noncomp_status_step.cc
-    test_noncomp_op_role.cc
+    test_noncomp_status_walk.cc
     test_noncomp_rewriter.cc
     test_noncomp_continuation.cc
     test_exact_driver.cc
