+++ tests/test_noncomp_validation.cc
@@ -26,9 +26,9 @@
 #include "clifft/noncomp/instrument_options.h"
 #include "clifft/noncomp/level.h"
 #include "clifft/noncomp/model.h"
-#include "clifft/noncomp/orchestrator.h"
 #include "clifft/noncomp/policy.h"
 #include "clifft/noncomp/rewriter.h"
+#include "clifft/noncomp/sample.h"
 #include "clifft/noncomp/transition_instrument.h"
 #include "clifft/optimizer/hir_pass_manager.h"
 #include "clifft/optimizer/pass_factory.h"
